vault: Manually track state for hidden mounts
Previously, Vaults is relying on gio's VolumeMonitor to monitor "user
interesting devices and volumes on the computer". As soon as the mounted
path contains a hidden folder, it's no longer 'user interesting',
according to the library developers, and thus disappears from the list
of mounted folders. Consequently, Vaults can't detect if the folder has
already been mounted or not and won't correctly unmount it as it
guesses to be unmounted, since there is no entry.

Fix this by introducing manually tracking of Vaults that have hidden
mount directories. For that, we use the previously used "proc-mounts"
crate and look for the Vault within all mounts. However, automatic
tracking of external unmounting or mounting is missing with this
approach.

Closes: https://github.com/mpobaschnig/Vaults/issues/45